README.md
The Rental management system is a web development application build using the HTML, CSS, JavaScript, jQuery, MYSQL and php. The application is used to perform and collect all the information of the tenants in a systematic and efficient manner. The RMS allow the user to stream their profiles, edit profile, electricity bill, rent bill at any time and any place while admin can view their tenants profile, electricity bill and rent bill by fill the name and room number of the specific tenant. The various CSS elements are used to enhance the interface of the application and make it more user-friendly, efficient and responsive.  
                                                            The application includes the login and logout features. Only resister users are able to login into the website. It replaces all the pen-paper work in maintaining records of the tenants in the diary with the online web-based system.